+++ PFE/PFE/Models/PIECEACHATECHEANCE.cs
@@ -6,7 +6,7 @@ namespace PFE.Models
         public int? PCAID { get; set; }
         public int? PEAID { get; set; }
         public DateTime PEADATE { get; set; }
-        public int? PEAMONTANT { get; set; }
+        public float? PEAMONTANT { get; set; }
         public int? PEATAUX { get; set; }
         public int? RGTID { get; set; }
         public string PEAISREGLE { get; set; }
