+++ PFE/PFE/Models/PIECEVENTETAXES.cs
@@ -3,15 +3,15 @@ namespace PFE.Models
 {
     public class PIECEVENTETAXES
     {
-        public int? PCVID { get; set; }
-        public int? CODETAXE { get; set; }
-        public int? PTVBASETVA { get; set; }
-        public int? PTVBASETVAESC { get; set; }
-        public int? PTVMNTTVA { get; set; }
-        public int? PTVTAUXTVA { get; set; }
-        public int? PTVBASETPF { get; set; }
-        public int? PTVBASETPFESC { get; set; }
-        public int? PTVMNTTPF { get; set; }
-        public int? PTVTAUXTPF { get; set; }
+        public float? PCVID { get; set; }
+        public float? CODETAXE { get; set; }
+        public float? PTVBASETVA { get; set; }
+        public float? PTVBASETVAESC { get; set; }
+        public float? PTVMNTTVA { get; set; }
+        public float? PTVTAUXTVA { get; set; }
+        public float? PTVBASETPF { get; set; }
+        public float? PTVBASETPFESC { get; set; }
+        public float? PTVMNTTPF { get; set; }
+        public float? PTVTAUXTPF { get; set; }
     }
 }
